feat: Lotto toString() 함수 추가
로또 출력용 함수 추가
